+++ src/gravity-client/NopGravityClient.ts
@@ -2,6 +2,7 @@ import {
   AddSessionRecordingResponse,
   AddSessionUserActionsResponse,
   IdentifySessionResponse,
+  ReadSessionCollectionSettingsResponse,
   SessionTraits,
   SessionUserAction,
 } from '../types'
@@ -25,5 +26,15 @@ export default class NopGravityClient extends AbstractGravityClient implements I
     screenRecords: ReadonlyArray<eventWithTime>,
   ): Promise<AddSessionRecordingResponse> {
     return { error: null }
+  }
+
+  async readSessionCollectionSettings(): Promise<ReadSessionCollectionSettingsResponse> {
+    return {
+      settings: {
+        sessionRecording: true,
+        videoRecording: false,
+      },
+      error: null,
+    }
   }
 }