+++ CHANGELOG.md
@@ -11,13 +11,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Option to disable video recording of sessions
-
 ### Changed
 
 ### Deprecated
 
 ### Removed
+
+- Option to disable video recording of sessions (now available from Gravity app)
 
 ### Fixed